+++ contrib/packages.json
@@ -7,7 +7,7 @@
         "github:nyoom-engineering/oxocarbon.nvim": {}
       }
     },
-    "git:github.com:catppuccin/nvim": {
+    "github.com:catppuccin/nvim": {
       "version": "tag:v0.2.9",
       "rename": "catppuccin",
       "build": "echo 'hi'"
